# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend for SkillFort.",
        "title": "SkillFort-backend API docs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/sf/v1",
    "paths": {
        "/admin/switch-branch": {
            "post": {
                "description": "Switch a record from one channel to another",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Switch a record from one channel to another",
                "parameters": [
                    {
                        "description": "batch",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.SwitchBranch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/channel": {
            "get": {
                "description": "Get all active channels",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "Get all active channels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Channel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DataResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Saves an instance of Channel",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "Save Channel",
                "parameters": [
                    {
                        "description": "channel",
                        "name": "channel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Channel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/channel/autocomplete": {
            "get": {
                "description": "Autocomplete for channels",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "Autocomplete for channels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.ChannelAutoComplete"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DataResponse"
                        }
                    }
                }
            }
        },
        "/channel/{id}": {
            "get": {
                "description": "Get an instance of Channel",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "Get a specific  Channel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Channel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DataResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an instance of Channel",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "Update Channel",
                "parameters": [
                    {
                        "description": "channel",
                        "name": "channel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Channel"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Channel id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an instance of Channel",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "Delete Channel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.Pagination": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "skipPagination": {
                    "type": "boolean"
                },
                "sort": {
                    "type": "string"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRecords": {
                    "type": "integer"
                }
            }
        },
...
        "errs.ErrorType": {
            "type": "string",
            "enum": [
                "OTHER",
                "INVALID",
                "IO",
                "EXIST",
                "NOT_EXIST",
                "PRIVATE",
                "INTERNAL",
                "BROKEN_LINK",
                "DATABASE",
                "VALIDATION",
                "UNANTICIPATED",
                "INVALID_REQUEST",
                "SMTP_ERROR",
                "EMAIL_ERROR",
                "STORAGE",
                "INSUFFICIENT_ACCESS",
                "CUSTOMER_DUPLICATE_TRANSACTION"
            ],
            "x-enum-comments": {
                "BROKEN_LINK": "Link target does not exist.",
                "CUSTOMER_DUPLICATE_TRANSACTION": "Duplicate Transaction",
                "DATABASE": "Error from database.",
                "EMAILERROR": "SMTP Error",
                "EXIST": "Item already exists.",
                "INSUFFICIENT_ACCESS": "Insufficient Access",
                "INTERNAL": "Internal error or inconsistency.",
                "INVALID": "Invalid operation for this type of item.",
                "INVALID_REQUEST": "Invalid Request",
                "IO": "External I/O error such as network failure.",
                "NOT_EXIST": "Item does not exist.",
                "OTHER": "Unclassified error. This value is not printed in the error message.",
                "PRIVATE": "Information withheld.",
                "SMTPERROR": "SMTP Error",
                "STORAGE": "Storage Error",
                "UNANTICIPATED": "Unanticipated error.",
                "VALIDATION": "Input validation error."
            },
            "x-enum-varnames": [
                "OTHER",
                "INVALID",
                "IO",
                "EXIST",
                "NOT_EXIST",
                "PRIVATE",
                "INTERNAL",
                "BROKEN_LINK",
                "DATABASE",
                "VALIDATION",
                "UNANTICIPATED",
                "INVALID_REQUEST",
                "SMTPERROR",
                "EMAILERROR",
                "STORAGE",
                "INSUFFICIENT_ACCESS",
                "CUSTOMER_DUPLICATE_TRANSACTION"
            ]
        },
        "errs.XError": {
            "type": "object",
            "properties": {
                "error": {},
                "errorCode": {
                    "type": "string"
                },
                "errorType": {
                    "$ref": "#/definitions/errs.ErrorType"
                },
                "params": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "requestModel.Channel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "ownerUserId": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "requestModel.SwitchBranch": {
            "type": "object",
            "properties": {
                "enquiryId": {
                    "type": "integer"
                },
                "studentId": {
                    "type": "integer"
                },
                "toChannel": {
                    "type": "integer"
                }
            }
        },
...
        "response.DataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/errs.XError"
                },
                "message": {
                    "type": "string"
                },
                "pageMetaData": {
                    "$ref": "#/definitions/db.Pagination"
                },
                "type": {
                    "$ref": "#/definitions/response.ResponseType"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/errs.XError"
                },
                "message": {
                    "type": "string"
                },
                "pageMetaData": {
                    "$ref": "#/definitions/db.Pagination"
                },
                "type": {
                    "$ref": "#/definitions/response.ResponseType"
                }
            }
        },
        "response.ResponseType": {
            "type": "string",
            "enum": [
                "Success",
                "Failure",
                "Warning"
            ],
            "x-enum-varnames": [
                "SUCCESS",
                "FAILURE",
                "WARNING"
            ]
        },
        "responseModel.Channel": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "channelId": {
                    "description": "Channel Id must be create only since it will interfere with update operations\nUse tx.Exec as raw query to update the channelId as done in channel after-create hook",
                    "type": "integer"
                },
                "channelOwnerFirstName": {
                    "type": "string"
                },
                "channelOwnerLastName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdById": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "ownerUserId": {
                    "description": "Reference",
                    "type": "integer"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedById": {
                    "type": "integer"
                }
            }
        },
        "responseModel.ChannelAutoComplete": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}